test: decouple docs routing policy from generated index
